:construction: (RUMM-399): Move RumEvent to dedicated package
