+++ dd-sdk-android/src/test/kotlin/com/datadog/android/rum/assertj/RumEventDataResourceAssert.kt
@@ -7,7 +7,7 @@
 package com.datadog.android.rum.assertj
 
 import com.datadog.android.rum.RumResourceKind
-import com.datadog.android.rum.internal.domain.RumEventData
+import com.datadog.android.rum.internal.domain.event.RumEventData
 import org.assertj.core.api.AbstractObjectAssert
 import org.assertj.core.api.Assertions.assertThat
