+++ dd-sdk-android/src/test/kotlin/com/datadog/android/rum/internal/domain/RumEventSerializerTest.kt
@@ -8,6 +8,9 @@ package com.datadog.android.rum.internal.domain
 
 import com.datadog.android.log.internal.user.UserInfo
 import com.datadog.android.rum.RumAttributes
+import com.datadog.android.rum.internal.domain.event.RumEvent
+import com.datadog.android.rum.internal.domain.event.RumEventData
+import com.datadog.android.rum.internal.domain.event.RumEventSerializer
 import com.datadog.android.utils.forge.Configurator
 import com.datadog.tools.unit.assertj.JsonObjectAssert.Companion.assertThat
 import com.datadog.tools.unit.extensions.ApiLevelExtension
