+++ dd-sdk-android/src/test/kotlin/com/datadog/android/utils/forge/RumEventDataUserActionForgeryFactory.kt
@@ -6,7 +6,7 @@
 
 package com.datadog.android.utils.forge
 
-import com.datadog.android.rum.internal.domain.RumEventData
+import com.datadog.android.rum.internal.domain.event.RumEventData
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.ForgeryFactory
